// src/app/cool-privilege-control/models/privilege-type-vm.ts
export class PrivilegeTypeVm {
    View: boolean = false;
    Search: boolean = false;
    Create: boolean = false;
    Update: boolean = false;
    Delete: boolean = false;
    Process: boolean = false;
    Generate: boolean = false;
    Import: boolean = false;
    Export: boolean = false;
    Login: boolean = false;
    Logout: boolean = false;
}